// src/pages/maintenance/types.d.ts
import { ImageCMS, TextCMS } from "@/types";

export interface TransportContent {
  banner: ImageCMS;
  firstAbout: ImageCMS;
  secondAbout: ImageCMS;
  thirdAbout: ImageCMS;
  fourthAbout: ImageCMS;
  fifthAbout: ImageCMS;
  map: ImageCMS;
  logo: ImageCMS;
  gridCards: ImageCMS[];
  simpleInformation: TextCMS;
}